requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.8.0
asyncpg==0.30.0
attrs==24.3.0
aws-sam-translator==1.94.0
aws-xray-sdk==2.14.0
bcrypt==4.0.1
billiard==4.2.1
black==24.10.0
boto3==1.35.95
botocore==1.35.95
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
cfn-lint==1.22.3
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.0
email_validator==2.2.0
-e git+https://github.com/oz-EndOfDay/BE.git@fbbf33dcd32cfc4a550b65c60d6c88c47883d9bf#egg=endofday
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-pagination==0.12.34
filelock==3.16.1
fsspec==2024.12.0
graphql-core==3.2.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joserfc==1.0.1
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-spec==0.1.3
jsonschema-specifications==2024.10.1
kombu==5.4.2
lazy-object-proxy==1.10.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moto==5.0.26
mpmath==1.3.0
multipart==1.2.1
mypy==1.14.1
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.1
openai==1.59.5
openapi-schema-validator==0.4.3
openapi-spec-validator==0.5.5
packaging==24.2
passlib==1.7.4
pathable==0.4.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
poetry-core==1.9.1
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-partiql-parser==0.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
responses==0.25.3
rfc3339-validator==0.1.4
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
xmltodict==0.14.2
